Take a first stab at porting my site
This has a generator including a server that re-renders on file changes,
a templating system using Tera, and Markdown rendering using
pulldown-cmark.

It's currently only rendering a single post, as well as the archive
page.

No feed yet, and no CV PDF generation yet.

CSS is also somewhat broken.

But it works.