Strip checkboxes from feed
I realised that the checkboxes used for the footnotes are included in
the feed content, and there is no CSS in that version, so they
potentially actually show up. As it stands, the footnote content is
still included in the feed, which is a bit weird, but it's a start.